Run the KMeans algorithm with different k
The results are saved in clustering_results.csv.
This file contain the k, silhuette score and the centers of the clusters.
It will be used later to select the optimal k.

The models with the corrensponding k are saved in tmp_models to avoid to
refit them once the optimal k is selected.